requirements.txt
Django==4.2.3
Pillow==10.0.0
django-admin-sortable2==2.1.8